+++ src/matrix/matrix.h
@@ -17,7 +17,7 @@
 # include "../structures.h"
 
 t_matrix	*identity_matrix(void);
-t_matrix	*new_matrix(int size, float initial_values[4][4]);
+t_matrix	*new_matrix(int size, double initial_values[4][4]);
 
 int			matrix_equals(t_matrix *a, t_matrix *b);
 int			is_invertible(t_matrix *a);
@@ -28,17 +28,17 @@ t_matrix	*matrix_multiply_n(t_matrix **transformations);
 t_tuple		*matrix_multiply_tuple(t_matrix *a, t_tuple *b);
 t_matrix	*transpose(t_matrix *matrix);
 
-float		determinant(t_matrix *a);
-float		minor(t_matrix *a, int row, int col);
-float		cofactor(t_matrix *a, int row, int col);
+double		determinant(t_matrix *a);
+double		minor(t_matrix *a, int row, int col);
+double		cofactor(t_matrix *a, int row, int col);
 t_matrix	*inverse(t_matrix *matrix);
 
 t_matrix	*submatrix(t_matrix *a, int row, int col);
 
-t_matrix	*translation(float x, float y, float z);
-t_matrix	*scaling(float x, float y, float z);
-t_matrix	*rotation_x(float rad);
-t_matrix	*rotation_y(float rad);
-t_matrix	*rotation_z(float rad);
+t_matrix	*translation(double x, double y, double z);
+t_matrix	*scaling(double x, double y, double z);
+t_matrix	*rotation_x(double rad);
+t_matrix	*rotation_y(double rad);
+t_matrix	*rotation_z(double rad);
 
 #endif /* !MATRIX_H */
